feat: 여행 목록 조회(GET /trips) 비로그인 접근 허용
- SecurityConfig: GET /trips 요청에 대해 permitAll() 설정 추가
- 프론트엔드 요청 반영: 비로그인 사용자도 여행 목록 페이지 접근 가능하도록 변경
- (참고: 국내/해외 카테고리 필터링 기능은 Map 작업 완료 후 추가 예정)

Resolves: 프론트요청사항 2번
