+++ build.gradle
@@ -29,6 +29,7 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation "org.springdoc:springdoc-openapi-starter-webmvc-api:2.8.5"
     implementation "org.springdoc:springdoc-openapi-starter-webmvc-ui:2.8.5"
+    implementation 'org.springframework.boot:spring-boot-starter-security'
     implementation 'org.springframework.security:spring-security-crypto'
 
     compileOnly 'org.projectlombok:lombok'
